condensed test files
